# api/openapi.yaml
openapi: 3.0.3
info:
  title: AI Job Orchestrator
  version: 0.1.0
  description: API contract for the in-memory MVP orchestrator.
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /jobs/{id}/trace:
    get:
      summary: Get trace timeline for a job
      parameters:
        - $ref: "#/components/parameters/JobID"
      responses:
        "200":
          description: Span timeline
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpanRecord"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    JobID:
      in: path
      name: id
      required: true
      schema:
        type: string
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    SpanRecord:
      type: object
      required: [name, timestamp, fields]
      properties:
        name:
          type: string
        timestamp:
          type: string
          format: date-time
        fields:
          type: object
          additionalProperties:
            type: string